+++ packages/program/src/functions/base-invocation-scope.ts
@@ -8,7 +8,8 @@ import { InputType } from '@fuel-ts/transactions';
 import { MAX_GAS_PER_TX } from '@fuel-ts/transactions/configs';
 import type { BaseWalletUnlocked } from '@fuel-ts/wallet';
 
-import { contractCallScript } from '../contract-call-script';
+import { getContractCallScript } from '../contract-call-script';
+import { POINTER_DATA_OFFSET } from '../script-request';
 import type { ContractCall, InvocationScopeLike, TransactionCostOptions, TxParams } from '../types';
 import { assert } from '../utils';
 
@@ -20,22 +21,12 @@ import { InvocationCallResult, FunctionInvocationResult } from './invocation-res
  * @param funcScope - The invocation scope containing the necessary information for the contract call.
  * @returns The contract call object.
  */
-function createContractCall(funcScope: InvocationScopeLike): ContractCall {
-  const { program, args, forward, func, callParameters, bytesOffset } = funcScope.getCallConfig();
-
-  const provider = program.provider as Provider;
-  const consensusParams = provider.consensusParamsCache;
-
-  if (!consensusParams) {
-    throw new Error(
-      'Provider consensus params cache empty! Pls make sure you ran `await Provider.connect()` and not just `new Provider()`'
-    );
-  }
-
-  const data = func.encodeArguments(
-    args as Array<InputValue>,
-    contractCallScript.getScriptDataOffset(consensusParams.maxInputs.toNumber()) + bytesOffset
-  );
+function createContractCall(funcScope: InvocationScopeLike, offset: number): ContractCall {
+  const { program, args, forward, func, callParameters } = funcScope.getCallConfig();
+  const DATA_POINTER_OFFSET = funcScope.getCallConfig().func.isInputDataPointer()
+    ? POINTER_DATA_OFFSET
+    : 0;
+  const data = func.encodeArguments(args as Array<InputValue>, offset + DATA_POINTER_OFFSET);
 
   return {
     contractId: (program as AbstractContract).id,
@@ -79,7 +70,18 @@ export class BaseInvocationScope<TReturn = any> {
    * @returns An array of contract calls.
    */
   protected get calls() {
-    return this.functionInvocationScopes.map((funcScope) => createContractCall(funcScope));
+    const script = getContractCallScript(this.functionInvocationScopes.length);
+    const provider = this.program.provider as Provider;
+    const consensusParams = provider.consensusParamsCache;
+    if (!consensusParams) {
+      throw new Error(
+        'Provider consensus params cache empty! Pls make sure you ran `await Provider.connect()` and not just `new Provider()`'
+      );
+    }
+    const maxInputs = consensusParams.maxInputs.toNumber();
+    return this.functionInvocationScopes.map((funcScope) =>
+      createContractCall(funcScope, script.getScriptDataOffset(maxInputs))
+    );
   }
 
   /**
@@ -90,6 +92,7 @@ export class BaseInvocationScope<TReturn = any> {
     calls.forEach((c) => {
       this.transactionRequest.addContractInputAndOutput(c.contractId);
     });
+    const contractCallScript = getContractCallScript(this.functionInvocationScopes.length);
     this.transactionRequest.setScript(contractCallScript, calls);
   }