+++ packages/program/src/script-request.ts
@@ -2,7 +2,14 @@
 import type { BytesLike } from '@ethersproject/bytes';
 import { arrayify } from '@ethersproject/bytes';
 import { Logger } from '@ethersproject/logger';
-import { TRANSACTION_SCRIPT_FIXED_SIZE, calculateVmTxMemory } from '@fuel-ts/abi-coder';
+import {
+  ASSET_ID_LEN,
+  CONTRACT_ID_LEN,
+  TRANSACTION_SCRIPT_FIXED_SIZE,
+  VM_TX_MEMORY,
+  WORD_SIZE,
+  calculateVmTxMemory,
+} from '@fuel-ts/abi-coder';
 import type { BN } from '@fuel-ts/math';
 import type {
   TransactionResultReturnDataReceipt,
@@ -14,13 +21,16 @@ import type {
   TransactionResult,
 } from '@fuel-ts/providers';
 import type { ReceiptScriptResult } from '@fuel-ts/transactions';
-import { ReceiptType, ByteArrayCoder } from '@fuel-ts/transactions';
+import { ReceiptType } from '@fuel-ts/transactions';
 import { versions } from '@fuel-ts/versions';
 
 import { ScriptResultDecoderError } from './errors';
 import type { CallConfig } from './types';
 
 const logger = new Logger(versions.FUELS);
+
+export const SCRIPT_DATA_BASE_OFFSET = VM_TX_MEMORY + TRANSACTION_SCRIPT_FIXED_SIZE;
+export const POINTER_DATA_OFFSET = ASSET_ID_LEN + 2 * WORD_SIZE + CONTRACT_ID_LEN + 2 * WORD_SIZE;
 
 /**
  * Represents a script result, containing information about the script execution.
@@ -94,7 +104,7 @@ function callResultToScriptResult(callResult: CallResult): ScriptResult {
  * @returns The decoded result.
  * @throws Throws an error if decoding fails.
  */
-function decodeCallResult<TResult>(
+export function decodeCallResult<TResult>(
   callResult: CallResult,
   decoder: (scriptResult: ScriptResult) => TResult,
   logs: Array<any> = []
@@ -163,6 +173,8 @@ export function callResultToInvocationResult<TReturn>(
   );
 }
 
+export type EncodedScriptCall = Uint8Array | { data: Uint8Array; script: Uint8Array };
+
 /**
  * `ScriptRequest` provides functionality to encode and decode script data and results.
  *
@@ -178,7 +190,7 @@ export class ScriptRequest<TData = void, TResult = void> {
   /**
    * A function to encode the script data.
    */
-  scriptDataEncoder: (data: TData) => Uint8Array;
+  scriptDataEncoder: (data: TData) => EncodedScriptCall;
 
   /**
    * A function to decode the script result.
@@ -194,7 +206,7 @@ export class ScriptRequest<TData = void, TResult = void> {
    */
   constructor(
     bytes: BytesLike,
-    scriptDataEncoder: (data: TData) => Uint8Array,
+    scriptDataEncoder: (data: TData) => EncodedScriptCall,
     scriptResultDecoder: (scriptResult: ScriptResult) => TResult
   ) {
     this.bytes = arrayify(bytes);
@@ -205,17 +217,15 @@ export class ScriptRequest<TData = void, TResult = void> {
   /**
    * Gets the script data offset for the given bytes.
    *
-   * @param bytes - The bytes of the script.
+   * @param byteLength - The byte length of the script.
    * @param maxInputs - The maxInputs value from the chain's consensus params.
    * @returns The script data offset.
    */
-  static getScriptDataOffsetWithBytes(bytes: Uint8Array, maxInputs: number): number {
-    const VM_TX_MEMORY = calculateVmTxMemory({ maxInputs });
-    return (
-      VM_TX_MEMORY + TRANSACTION_SCRIPT_FIXED_SIZE + new ByteArrayCoder(bytes.length).encodedLength
-    );
+  static getScriptDataOffsetWithScriptBytes(byteLength: number, maxInputs: number): number {
+    const scriptDataBaseOffset = calculateVmTxMemory({ maxInputs }) + TRANSACTION_SCRIPT_FIXED_SIZE;
+    return scriptDataBaseOffset + byteLength;
   }
-  
+
   /**
    * Gets the script data offset.
    *
@@ -223,7 +233,7 @@ export class ScriptRequest<TData = void, TResult = void> {
    * @returns The script data offset.
    */
   getScriptDataOffset(maxInputs: number) {
-    return ScriptRequest.getScriptDataOffsetWithBytes(this.bytes, maxInputs);
+    return ScriptRequest.getScriptDataOffsetWithScriptBytes(this.bytes.length, maxInputs);
   }
 
   /**
@@ -243,7 +253,15 @@ export class ScriptRequest<TData = void, TResult = void> {
    * @returns The encoded data.
    */
   encodeScriptData(data: TData): Uint8Array {
-    return this.scriptDataEncoder(data);
+    const callScript = this.scriptDataEncoder(data);
+    // if Uint8Array
+    if (ArrayBuffer.isView(callScript)) {
+      return callScript;
+    }
+
+    // object
+    this.bytes = arrayify(callScript.script);
+    return callScript.data;
   }
 
   /**