+++ packages/providers/src/transaction-request/script-transaction-request.ts
@@ -117,8 +117,8 @@ export class ScriptTransactionRequest extends BaseTransactionRequest {
    * @param data - The script data.
    */
   setScript<T>(script: AbstractScriptRequest<T>, data: T) {
-    this.script = script.bytes;
     this.scriptData = script.encodeScriptData(data);
+    this.script = script.bytes;
 
     if (this.bytesOffset === undefined) {
       this.bytesOffset = this.scriptData.byteLength;