+++ packages/providers/src/transaction-summary/assemble-transaction-summary.ts
@@ -1,6 +1,6 @@
 import { hexlify } from '@ethersproject/bytes';
 import type { BN } from '@fuel-ts/math';
-import type { Transaction } from '@fuel-ts/transactions';
+import { type Transaction } from '@fuel-ts/transactions';
 
 import type { TransactionResultReceipt } from '../transaction-response';
 import { calculateTransactionFee } from '../utils';
@@ -12,6 +12,7 @@ import {
   isTypeCreate,
   isTypeScript,
 } from './operations';
+import { extractBurnedAssetsFromReceipts, extractMintedAssetsFromReceipts } from './receipt';
 import { processGraphqlStatus } from './status';
 import type { AbiParam, GraphqlTransactionStatus, TransactionSummary } from './types';
 
@@ -43,14 +44,14 @@ export function assembleTransactionSummary<TTransactionType = void>(
     abiParam,
   } = params;
 
-  const { gasUsed, fee } = calculateTransactionFee({
-    receipts,
+  const { fee, gasUsed } = calculateTransactionFee({
     gasPrice,
+    transactionBytes,
+    transactionWitnesses: transaction?.witnesses || [],
     gasPerByte,
     gasPriceFactor,
-    transactionBytes,
     transactionType: transaction.type,
-    transactionWitnesses: transaction?.witnesses || [],
+    receipts,
   });
 
   const operations = getOperations({
@@ -67,6 +68,9 @@ export function assembleTransactionSummary<TTransactionType = void>(
   const { isStatusFailure, isStatusPending, isStatusSuccess, blockId, status, time } =
     processGraphqlStatus(gqlTransactionStatus);
 
+  const mintedAssets = extractMintedAssetsFromReceipts(receipts);
+  const burnedAssets = extractBurnedAssetsFromReceipts(receipts);
+
   const transactionSummary: TransactionSummary<TTransactionType> = {
     id,
     fee,
@@ -77,6 +81,8 @@ export function assembleTransactionSummary<TTransactionType = void>(
     time,
     status,
     receipts,
+    mintedAssets,
+    burnedAssets,
     isTypeMint: isTypeMint(transaction.type),
     isTypeCreate: isTypeCreate(transaction.type),
     isTypeScript: isTypeScript(transaction.type),